NetBox: MoreMessageDialog etc

+++ src/NetBox/FarPlugin.h
@@ -585,7 +585,7 @@ public:
     TStrings *History, bool PathInput,
     TInputDialogInitializeEvent OnInitialize, bool Echo) override;
   uint32_t MoreMessageDialog(const UnicodeString AMessage,
-    TStrings *MoreMessages, TQueryType Type, uintptr_t Answers,
+    TStrings *MoreMessages, TQueryType Type, uint32_t Answers,
     const TMessageParams *Params) override;
 };
